change Float to Integer

+++ extraCredit/AddScore.java
@@ -3,23 +3,23 @@ package extraCredit;
 import java.util.*;
 import java.util.Comparator;
 public class AddScore{
-	public float score;
-	public ArrayList<ArrayList<Float>> table;
-	public ArrayList<ArrayList<Float>> tableWithScore;
+	public Integer score;
+	public ArrayList<ArrayList<Integer>> table;
+	public ArrayList<ArrayList<Integer>> tableWithScore;
 	public int[] v;
-	public AddScore(ArrayList<ArrayList<Float>> table, int[] v){
+	public AddScore(ArrayList<ArrayList<Integer>> table, int[] v){
 		this.table = table;
 		this.v = v;
 		tableWithScore = ScoreAdd();
 	}
-	//private float score;
+	//private Integer score;
 	//caculate the score of each tuple and add to each tuple
 	
-	public  ArrayList<ArrayList<Float>> ScoreAdd(){
-		ArrayList<ArrayList<Float>> tableWithScore = new ArrayList<ArrayList<Float>>();
-		for (ArrayList<Float> tuple : table){
-			 score = 0.00f;
-			 ArrayList<Float> tupleWithScore = new ArrayList<Float>();
+	public  ArrayList<ArrayList<Integer>> ScoreAdd(){
+		ArrayList<ArrayList<Integer>> tableWithScore = new ArrayList<ArrayList<Integer>>();
+		for (ArrayList<Integer> tuple : table){
+			 score = 0;
+			 ArrayList<Integer> tupleWithScore = new ArrayList<Integer>();
 			 tupleWithScore.addAll(tuple);
 			for (int i = 0; i < v.length;i++)
 			{
@@ -31,9 +31,9 @@ public class AddScore{
 			tableWithScore.add(tupleWithScore);
 		}
 		//descending order
-		Collections.sort(tableWithScore, new Comparator<ArrayList<Float>> () {
+		Collections.sort(tableWithScore, new Comparator<ArrayList<Integer>> () {
 		    @Override
-		    public int compare(ArrayList<Float> a, ArrayList<Float> b) {
+		    public int compare(ArrayList<Integer> a, ArrayList<Integer> b) {
 		        return b.get(v.length).compareTo(a.get(v.length));
 		    }
 		});
